Warning cleanups. Lots of -no-warns still turned on.

+++ Language/KansasLava/Testing/Observe.hs
@@ -14,6 +14,6 @@ observeRep :: forall a. (Rep a) => String -> Seq a -> Seq a
 observeRep msg = dual shallow id
   where shallow
 	  = shallowSeq
-	  . foldr (\ (i,x) xs -> trace (msg ++ "(" ++ show i ++ ")" ++ showRep (Witness :: Witness a) x) $ x :~ xs) (error "never done") 
-	  . zip [0..]
+	  . foldr (\ (i,x) xs -> trace (msg ++ "(" ++ show i ++ ")" ++ showRep (Witness :: Witness a) x) $ x :~ xs) (error "never done")
+	  . zip [(0::Int)..]
 	  . fromSeqX
